Moving around python scripts
