+++ scripts/monitor.py
@@ -5,8 +5,8 @@ import logging
 import argparse
 import sys
 
-from utils.igrill import IGrillPeripheral, IGrillMiniPeripheral
-from utils.utils import SetupLog
+from py_utils.igrill import IGrillPeripheral, IGrillMiniPeripheral
+from py_utils.logging import SetupLog
 from py_config.mac_config import ADDRESS
 
 DATA_FILE = sys.path[0]+'/../run/igrill.json'
